requirements.txt
appdirs==1.4.4
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
cycler==0.10.0
docker-pycreds==0.4.0
filelock==3.12.2
fonttools==4.42.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
huggingface-hub==0.16.4
idna==2.10
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.24.0
psutil==5.9.5
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
RandAugment==0.1
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.2
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.1
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
supervision==0.13.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchcam==0.3.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wandb==0.15.8
wget==3.2
